Port Trilinos wrappers from Epetra/AztecOO to Tpetra/Belos

See merge request DAER/mbdyn!257

+++ mbdyn/base/nonlin.h
@@ -304,7 +304,7 @@ struct CommonNonlinearSolverParam
           PRINT_CONVERGENCE_INFO = 0x2
      };
      
-     CommonNonlinearSolverParam(unsigned uFlags_a = 0u,
+     CommonNonlinearSolverParam(long unsigned uFlags_a = 0u,
                                 integer iIterBeforeAss = 0,
                                 bool bKeepJac = false)
           :uFlags(uFlags_a),
@@ -312,7 +312,7 @@ struct CommonNonlinearSolverParam
            bKeepJacAcrossSteps(bKeepJac) {
      }
 
-     unsigned uFlags;
+     long unsigned uFlags;
      integer iIterationsBeforeAssembly;
      bool bKeepJacAcrossSteps;     
 };
@@ -525,16 +525,16 @@ public:
 		return pSolTest;
 	}
 
-        enum NonlinearSolverHintReal {
-                LINESEARCH_LAMBDA_MAX,
-                LINESEARCH_LAMBDA_CURR,
-                NONLINEAR_SOLVER_LAST_HINT_REAL
-        };
+	enum NonlinearSolverHintReal {
+			LINESEARCH_LAMBDA_MAX,
+			LINESEARCH_LAMBDA_CURR,
+			NONLINEAR_SOLVER_LAST_HINT_REAL
+	};
 
-        enum NonlinearSolverHintInteger {
-                LINESEARCH_ITERATION_CURR,
-                NONLINEAR_SOLVER_LAST_HINT_INTEGER
-        };
+	enum NonlinearSolverHintInteger {
+			LINESEARCH_ITERATION_CURR,
+			NONLINEAR_SOLVER_LAST_HINT_INTEGER
+	};
 
 
     void SetNonlinearSolverHint(NonlinearSolverHintReal eType, doublereal dHint) {
